Data Table Update
Added in all of my code for in the first view controller for the table
management. There may need to be copying of the X and Y value arrays
(inputXValues and inputYValues) when the view is switched so that math
can be done on them, but I haven’t reviewed your most recent code yet
so I don’t know what to do with them.