Bug Fixes/Minor Tweaks
Fixed a bug that crashes the plot when the exponential function grows
too quickly for the window. Made it so the limits for the graph are
placed in the middle of the screen if the axis is off the screen. Also
shrunk down the display for the trendline equation to %.2f values
instead of %f values since users can get more precise values from the
previous tab. Also removed an extra pair of parentheses in the text
display of the power regression r value and trendline.